# Improve performance - Avoid repeated git diff calls